Removing wrangler old reparent code.
